sdr-agent/pyproject.toml
[project]
name = "sdr-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "anthropic>=0.40.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.28.0",
    "redis>=5.2.0",
    "psycopg2-binary>=2.9.10",
    "psycopg[binary]>=3.2.0",
    "psycopg-pool>=3.2.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.2",
    "tavily-python>=0.5.0",
    "tenacity>=9.0.0",
    "structlog>=24.4.0",
    "langsmith>=0.2.0",
    "asyncpg>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]
